backend-send-config-commands/requirements.txt
firebase-admin==2.9.1
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
